move v2 refactored tests to separate file, remove TestSameBlock as it is already covered in TransactionChainIndices refactored tests
